backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pillow==11.0.0
numpy==2.1.0
rapidocr-onnxruntime>=1.3.19
openai==2.8.1
langsmith>=0.1.0
sentry-sdk[fastapi]>=2.0.0
pydantic==2.9.0
